Move stoppable thread with server to testing tools
