+++ conans/test/utils/tools.py
@@ -2,11 +2,13 @@ import os
 import shlex
 import shutil
 import sys
+import threading
 import uuid
 from collections import Counter
 from contextlib import contextmanager
 from io import StringIO
 
+import bottle
 import requests
 import six
 from mock import Mock
@@ -507,3 +509,19 @@ class TestClient(object):
         save_files(path, files)
         if not files:
             mkdir(self.current_folder)
+
+
+class StoppableThreadBottle(threading.Thread):
+
+    # Instance server and run it in a stoppable thread
+    server = None
+
+    def __init__(self, host="0.0.0.0", port=8080):
+        self.server = bottle.Bottle()
+        super(StoppableThreadBottle, self).__init__(target=self.server.run, kwargs={"host": host,
+                                                                                    "port": port})
+        self.daemon = True
+        self._stop = threading.Event()
+
+    def stop(self):
+        self._stop.set()
